Solan/backend/requirements.txt
Django>=3.0
djangorestframework>=3.11.0
Pillow>=7.0.0
